requirements.txt
django
sorl-thumbnail
django-allauth
Pillow
